Changes to CSS and HTML files
Added description, keywords, and icon meta tags in head of the HTML file.
Changed a bunch of the heading tags to divs. Also Changes a lot of id's to classes. Deleted a lot of divs. All done to make things easier to read and navigate. Also formatted the document for the same reason.

Made ground up changes to how the website works. Added grids to nav bar photo , and imbedded video sections.